# lib/api-zod/src/generated/types/knowledgeSectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgeSectionInput {
  title: string;
  data: unknown;
}